Move SSH PTY resize monitor into actor
