+++ .github/swift-file-length-budget.tsv
@@ -1,7 +1,7 @@
 # cmux-owned Swift file length budget.
 # Format: max_lines<TAB>relative path
 # Reduce counts as files shrink. CI fails if tracked files exceed this budget.
-34321	CLI/cmux.swift
+34147	CLI/cmux.swift
 17606	Sources/AppDelegate.swift
 16025	Sources/ContentView.swift
 14100	Sources/TerminalController.swift
